playground/opentelemetry/requirements-dev.txt
python-dotenv==1.0.1
Flask==3.0.2
requests==2.31.0

# Local Observability
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0

#
opentelemetry-instrumentation-flask==0.45b0
opentelemetry-instrumentation-requests==0.45b0